+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [1.0.0 - 2025-03-25]
+## [1.0.0 - 2025-11-25]
 ### Added
 - First version of the project
 - Spring Application
@@ -219,6 +219,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Check message for HTTP requests to exporter
 - Request Variable annotation
 - Optimize JPA configuration
+- Compatibility with Authentik
+- JWT Decoder configuration
 
 ### Changed
 - Rename project state CREATED and ACCEPTED as REVIEW and APPROVAL
+- Update to Maven 4
+- Update to Java 25
+
+### Removed
+- RStudio Keycloak Groups